requirements.txt
# Data Collection & API
requests==2.32.3
python-dotenv==1.0.1

# Kafka Integration
kafka-python==2.0.2
confluent-kafka==2.3.0

# Data Processing & Storage
pyspark==3.5.1
pyiceberg==0.7.1

# MongoDB Integration
pymongo==4.6.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.1
pandas==2.0.3
